reoganize functions
